IMP - Refactored component for image in MD
