+++ src/components/elements/markdown/components/link.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 
-import {ImageInHTML, isImage} from '../transformers/image';
+import {ImageInHTML} from '../transformers/image/main';
+import {isImage} from '../transformers/image/utils';
 import {EmbeddedYoutubeVideo, extractVideoId} from '../transformers/youtube';
 import {MarkdownComponentProps} from '../types';
 
@@ -16,7 +17,7 @@ export const renderLink = ({children, href, ...props}: LinkComponentProps) => {
   }
 
   if (isImage(href)) {
-    return <ImageInHTML imageUrl={href} alt="image"/>;
+    return <ImageInHTML src={href} alt="image"/>;
   }
 
   return <a href={href} target="_blank" rel="noreferrer" {...props}>{children}</a>;
