+++ src/components/elements/markdown/transformers/image/const.ts
@@ -1,0 +1,1 @@
+export const IMAGE_CLASS_SPLITTER = '%7C'; // Escaped "|"
