+++ src/components/elements/markdown/transformers/image/types.ts
@@ -1,0 +1,5 @@
+export type ImageProps = {
+  src: string,
+  alt: string,
+  className?: string,
+}
