+++ src/components/elements/markdown/transformers/image/utils.ts
@@ -1,0 +1,6 @@
+export const isImage = (imageUrl: string) => {
+  return imageUrl.endsWith('.jpeg') ||
+    imageUrl.endsWith('.png') ||
+    imageUrl.endsWith('.jpg') ||
+    imageUrl.endsWith('.gif');
+};
